requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
Faker==24.9.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.2.1
